Create KMP- Algorithm to search for a string pattern
KMP Algo is one of the famous algo to search for a string pattern from a string. It is better than basic algo "Naive algo" since it's complexity is O(mn) , whereas the this algo has complexity O(n+m), where  n and m are the string sizes. 
This Algorithm requires to compute the LPS(Longest Proper Prefix  Suffix) first.