Create Strings Rotations of Each Other.cpp
This question is about checking if the other string is of rotated string of first one or not.
I am taking help of KMP algorithm for searching algorithm.
Code explanation:
first concatenated the first string with itself. it will add the back part and front part together. So, if other string is ratated string of first one, then it will be present there.
Now using the KMP algo to search the pattern in the concatenated string and stored the output in a variable.
I added -1 in the back just to check if there is any element .
Now conditional statement will specify if there is any element or not.